requirements/base.txt
arrow==0.7.0
beautifulsoup4==4.4.1
Django==1.11.4
django-disqus==0.5
django-emoticons==1.2
djangorestframework==3.3
django-forms-bootstrap==3.1.0
django-tinymce==2.6.0
Pillow==3.1.1
python-dateutil==2.4.2
six==1.10.0
wheel==0.24.0